No need for so many casts
git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26192 bd178f5f-7bea-4014-844f-495ca19b8805

+++ src/org/starexec/data/database/Jobs.java
@@ -2583,10 +2583,10 @@ public class Jobs {
 				liteJobPairStats.put("pendingPairs", results.getInt("pendingPairs"));
 				liteJobPairStats.put("errorPairs", results.getInt("errorPairs"));
 
-				Integer completionPercentage = Math.round(100 * (float) (results.getInt("completePairs")) / ((float) results.getInt("totalPairs")));
+				Integer completionPercentage = Math.round(100f * results.getInt("completePairs") / results.getInt("totalPairs"));
 				liteJobPairStats.put("completionPercentage", completionPercentage);
 
-				Integer errorPercentage = Math.round(100 * (float) (results.getInt("errorPairs")) / ((float) results.getInt("totalPairs")));
+				Integer errorPercentage = Math.round(100f * results.getInt("errorPairs") / results.getInt("totalPairs"));
 				liteJobPairStats.put("errorPercentage", errorPercentage);
 
 				Job j = new Job();
